requirements.txt
# ── Voice Agent: Google ADK + Deepgram + LiveKit ──────────────────────────────
# Install with:  pip install -r requirements.txt

# ── LiveKit Agents (core + Deepgram + Silero VAD) ─────────────────────────────
livekit-agents>=1.4.0,<2.0.0
livekit-plugins-deepgram>=0.9.0
livekit-plugins-silero>=0.8.0
livekit                          # base LiveKit Python SDK
livekit-api                      # for token generation in Streamlit

# ── Google ADK + Gemini ───────────────────────────────────────────────────────
google-adk>=0.5.0
google-generativeai>=0.8.0       # underlying Gemini SDK used by ADK

# ── Streamlit UI ──────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
aiohttp>=3.9.0
